+++ CinemaManagement/BusinessObject/User.cs
@@ -17,8 +17,8 @@ namespace BusinessObject
 
 		[Key]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public int UserId { get; set; }
-		public int RoleId { get; set; }
+		public long UserId { get; set; }
+		public long RoleId { get; set; }
 		public string? Email { get; set; }
 		public string? Password { get; set; }
 		public string? FirstName { get; set; }
